+++ src/UI/OutputView.js
@@ -7,7 +7,10 @@ const OutputView = {
    * <p>
    * 출력을 위해 필요한 메서드의 인자(parameter)는 자유롭게 추가하거나 변경할 수 있다.
    */
-  printMap() {},
+  printMap(bridge, state, bridgeMaterialize) {
+    let [top, bottom] = bridgeMaterialize([], [], bridge);
+    top = top.
+  },
 
   /**
    * 게임의 최종 결과를 정해진 형식에 맞춰 출력한다.
